Add a bootstrap build option; enable bootstrap
ldc 1.0.0 and newer need a working D compiler to build, causing a
circular dependency on itself. To make it possible to bootstrap ldc,
upstream is going to keep on supporting 0.17.x for building newer ldc
versions as this is the last version that only needs a C++ compiler for
building.

This commit adds a bootstrap define to allow boostrapping with ldc
version 0.17.2.