Merge -devel subpackages into the main ldc package
ldc is a compiler and as such is already a development package, so
it doesn't make sense to split out separate -devel subpackages. The
development headers are only useful together with ldc, and ldc already
hard-depended on the split-out subpackages.

This commit drops the separate subpackages to make things simpler.