Remove geany tags subpackage
Pulling in geany as a build dep makes the buildroot huge as it has to
pull in all the graphical libraries for supporting geany.

Also, I strongly suspect nobody uses this. geany users are probably used
to generating the tags themselves for the things that they need. The
only packages currently in Fedora that ship geany tags are ldc and gtkd,
and nothing else.